front/pyproject.toml
[tool.poetry]
name = "front"
version = "0.1.0"
description = ""
authors = ["mielesantos <mielesnts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
